uart: extract btstack_uart_posix_block implementation
